Separa logica de UI en Login
