--- src/components/login/Login.js
@@ -1,62 +1,0 @@
-// src/components/Login.js
-import React, { useState } from "react";
-import { auth } from "../../firebaseConfig";
-import { signInWithEmailAndPassword } from "firebase/auth";
-import { useNavigate } from "react-router-dom";
-import "../Login.css";
-
-const Login = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const navigate = useNavigate();
-
-  const handleLogin = async (e) => {
-    e.preventDefault();
-    try {
-      await signInWithEmailAndPassword(auth, email, password);
-      navigate("/tasks");
-    } catch (error) {
-      console.error("Error al iniciar sesión:", error.message);
-    }
-  };
-
-  return (
-    <div className="login-container">
-      <div className="login-box">
-        <img
-          src="/checklist_small.png" 
-          alt="Login"
-          className="login-image"
-        />
-        <form onSubmit={handleLogin}>
-          <h2>Inicia Sesión</h2>
-          <input
-            type="email"
-            placeholder="Correo electrónico"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            className="login-input"
-          />
-          <input
-            type="password"
-            placeholder="Contraseña"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            className="login-input"
-          />
-          <button type="submit" className="login-button">
-            Iniciar Sesión
-          </button>
-        </form>
-        <p
-          onClick={() => navigate("/signup")}
-          className="toggle-auth"
-        >
-          ¿No tienes cuenta? Regístrate
-        </p>
-      </div>
-    </div>
-  );
-};
-
-export default Login;
